small changes to v2
I just read over V2 and made some slight edits in wording to all the sections. I adjusted wording in results section so that the we were consistent in saying that there was a statistically significant difference in finale vs non-finale. 
(I can't get the .qmd to render to the pdf agh sorry!)